Add documentation based on Sphinx
The documentation can be generated by running `make docs` in the repository.
Output files can be found in the `docs/_build` directory.
The documentation currently reflects the contents of `README.md`.
Docstrings that use markdown are correctly rendered in the API documentation.
Pandoc and the pypandoc package need to be installed for this to work.

PiperOrigin-RevId: 159215223