Make mask argument to Conv2D more flexible
The mask argument to Conv2D needed to be something that is convertible to a
numpy array. The numpy mask was then automatically converted to a TensorFlow
constant when applied.
Because the constants are stored in the graph, this increases the size of graph
quite considerably (by the size of the Conv2D filters).
If we allow the user to pass in a TensorFlow Tensor instead, it is easily
possible to use a PyFunc wrapper around a numpy array, or to cache constants
in between calls to Conv2D.
This extends Conv2D to be able to handle any mask input that is convertible
to a float32 or float64 Tensor.

PiperOrigin-RevId: 173877530