Slice TensorShape instead of Tensor in basic_rnn.py.
This prevents DeepRNN.output_size() from creating a slice operation on
a Tensor declared inside the dynamic RNN loop body (output_size() is
called outside the loop body, so the resulting slice operation is
invalid).  This currently works because the slice operation isn't
actually run (since it's only used to get its shape), but in the
future creating operations with invalid inputs will raise an
exception.

PiperOrigin-RevId: 177453503