Creation of a ConvND class. More flexible masks. Lots more tests.
Creating a ConvND class that is then subclassed for Conv{1,2,3}D. This removes a lot of duplicated boilerplate code. As part of this we, are also adding a `mask` argument to the Conv{1,3}D classes. Finally, we remove rank restrictions on the mask argument. We add testing for new masking functionality.

PiperOrigin-RevId: 179918026