Switch to using private member variables when possible.
There can be subtle differences to the public and private versions of member variables in the conv module (eg, Convolution's `stride` is the conv size + 2 but `_stride` is the conv size). The current module mixes and matches these. To reduce cognitive load a bit, we switch to using the private member variable when possible.

PiperOrigin-RevId: 198857690