Implement same graph check using graph keys.
tl;dr - resolves the last known issues with Sonnet and `tfe.defun`.

Inside a `defun` we observe a different graph instance each time the function is
traced. Sonnet asserts that each time a module is traced that it's graph has not
changed. This CL changes that test to check that the graph we observe each time
has the same "key" (rather than being the same instance). The key is akin to a
primary key for the graph. Graph instances with the same key form part of the
same parent graph (e.g. a sub-graph representing a function with key "a" shares
variables with a regular tf.Graph with key "a" which it is a part of).

DifferentGraphError used to fire undesirably in defun in the following cases
(both of which are fixed in this cl):

  1) `enter_variable_scope` is used in conjunction with `_build` (e.g. in the
     constructor or via `snt.reuse_variables`.
  2) `defun` re-traces `_build` due to the input signature changing (e.g.
     Tensor shape changing, or Python parameter values changing).

PiperOrigin-RevId: 209131983