ConvNet2D{Transpose}: Filter the kwargs passed to the normalizer.
This makes it possible to connect the module with is_training=True, even if the
normalization module selected is something which does not support that (eg
LayerNorm). Normal practice would be to call the module with is_training inside
a try block, catch any errors and then reconnect without is_training. However,
that will generally have created some variables internally, so global
(tf.Graph-level) state has been changed, producing errors.

This solution works for any normalization module which has a signature that
does not contain **kwargs, because that makes it impossible to introspect over
whether the flag is supported or not.

PiperOrigin-RevId: 224994092