Change _specialize_per_device to special-case eager mode
In eager mode a specialization is chosen based on the current device
scope, whereas in graph (tf.function) that choice is delegated to the
implementation selector pass in Grappler.

Handling the two cases differently is needed as currently the implementation
selector seems to itgnore tf.device(...) context outside of tf.function,
i.e.

with tf.device("GPU:0"):
  _specialized_fn(...)

will not be executed on "GPU:0", whereas

def _specialized_fn(...):
  with tf.device("GPU:0"):
    ...

_specialized_fn(...)

will be.

PiperOrigin-RevId: 259895315
Change-Id: I4a1e940f4d84235db4695691347ec31b37a0897e