Add Support for Mixed Precision Training
Mixed precision speeds up training by casting
layers to lower precision formats when it is numerically
stable to do so. For example, this API can be used to cast
to float16 for performance increases when training on GPUs.

This is implemented by casting inputs and weights of layers
marked eligible for mixed precision, and scaling losses if
necessary to ensure that gradients do not underflow to zero.

PiperOrigin-RevId: 264644827
Change-Id: Id433bbadcafb0d2ebced0d2ffa88fb103ef37cf1